# src/swagger_documentation.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "DigitalStore API",
    "description": "Documentação da API da digital store"
  },
  "host": "https://digital-store-backend-c4an.onrender.com",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "JWT": {
      "description": "JWT token",
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/banners/": {
      "get": {
        "tags": [
          "Banner"
        ],
        "description": "Retorna lista de banner",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna lista de banner",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "banner_id": {
                    "type": "string",
                    "example": "id do banner"
                  },
                  "banner_nome": {
                    "type": "string",
                    "example": "nome do banner"
                  },
                  "banner_subtitulo": {
                    "type": "string",
                    "example": "subtitulo"
                  },
                  "banner_titulo": {
                    "type": "string",
                    "example": "titulo"
                  },
                  "banner_descricao": {
                    "type": "string",
                    "example": "descrição"
                  },
                  "banner_link": {
                    "type": "string",
                    "example": "link do botão"
                  },
                  "banner_imagem": {
                    "type": "string",
                    "example": "binario da imagem"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Banner"
        ],
        "description": "Cria uma Banner",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "banner_nome": {
                  "type": "string",
                  "example": "nome do banner"
                },
                "banner_subtitulo": {
                  "type": "string",
                  "example": "subtitulo"
                },
                "banner_titulo": {
                  "type": "string",
                  "example": "titulo"
                },
                "banner_descricao": {
                  "type": "string",
                  "example": "descrição"
                },
                "banner_link": {
                  "type": "string",
                  "example": "link do botão"
                },
                "banner_imagem": {
                  "type": "string",
                  "example": "binario da imagem"
                }
              },
              "required": [
                "banner_nome",
                "banner_subtitulo",
                "banner_titulo",
                "banner_descricao",
                "banner_link",
                "banner_imagem"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Banner criado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "detail": {
                  "type": "string",
                  "example": "Banner criado"
                },
                "severity": {
                  "type": "string",
                  "example": "success"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Banner"
        ],
        "description": "Edita uma banner",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "banner_id": {
                  "type": "string",
                  "example": "id do banner"
                },
                "banner_nome": {
                  "type": "string",
                  "example": "nome do banner"
                },
                "banner_subtitulo": {
                  "type": "string",
                  "example": "subtitulo"
                },
                "banner_titulo": {
                  "type": "string",
                  "example": "titulo"
                },
                "banner_descricao": {
                  "type": "string",
                  "example": "descrição"
                },
                "banner_link": {
                  "type": "string",
                  "example": "link do botão"
                },
                "banner_imagem": {
                  "type": "string",
                  "example": "binario da imagem"
                }
              },
              "required": [
                "banner_id",
                "banner_nome",
                "banner_subtitulo",
                "banner_titulo",
                "banner_descricao",
                "banner_link",
                "banner_imagem"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Banner atualizado.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "detail": {
                  "type": "string",
                  "example": "Banner atualizado."
                },
                "severity": {
                  "type": "string",
                  "example": "success"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/banners/{id}": {
      "get": {
        "tags": [
          "Banner"
        ],
        "description": "Retorna um banner",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna um banner",
            "schema": {
              "type": "object",
              "properties": {
                "banner_id": {
                  "type": "string",
                  "example": "id do banner"
                },
                "banner_nome": {
                  "type": "string",
                  "example": "nome do banner"
                },
                "banner_subtitulo": {
                  "type": "string",
                  "example": "subtitulo"
                },
                "banner_titulo": {
                  "type": "string",
                  "example": "titulo"
                },
                "banner_descricao": {
                  "type": "string",
                  "example": "descrição"
                },
                "banner_link": {
                  "type": "string",
                  "example": "link do botão"
                },
                "banner_imagem": {
                  "type": "string",
                  "example": "binario da imagem"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Banner"
        ],
        "description": "Deleta um banner.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Banner deletado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "detail": {
                  "type": "string",
                  "example": "Banner deletado com sucesso."
                },
                "severity": {
                  "type": "string",
                  "example": "success"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
...
}